move threads implementation header stuff
