+++ ENIGMAsystem/SHELL/Platforms/General/PFthreads.cpp
@@ -1,3 +1,4 @@
+#include "PFthreads_impl.h"
 #include "PFthreads.h"
 
 #include "Universal_System/resource_data.h" //script_execute
