+++ ENIGMAsystem/SHELL/Platforms/Win32/include.h
@@ -1,5 +1,4 @@
-#include "WINDOWSmain.h"
-#include "../General/PFjoystick.h"  
+#include "../General/PFjoystick.h"
 #include "../General/PFthreads.h"
 #include "../General/PFini.h"
 #include "../General/PFfilemanip.h"
